code suggestions and basic outside
fixed some code that godot suggested, just renaming parameters that share names with another.

Also added some function to the outside scene, people move from state to state.

Todo:
the queue into the store and maybe one for the advertiser?, customer conversion,  signal that tells the game when a person enters a shop.

Need to modify the way move speed is handled, modify customer wait times for an advertiser